pyproject.toml
[project]
name = "aiush"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "fastapi",
    "uvicorn",
    "uvloop",
    "transformers",
    "torch==2.4.0+cu124",
    "trl",
    "bitsandbytes>=0.43.0",
    "loralib",
    "wandb",
    "accelerate>=0.33.0",
    "peft>=0.12.0",
    "huggingface-hub[cli]>=0.24.2",
    "discord-py>=2.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"    


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/aiush"]
